pyproject.toml
[project]
name = "uber-data-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.179.0",
    "google-cloud-bigquery>=3.36.0",
    "load-dotenv>=0.1.0",
    "loguru>=0.7.3",
    "pre-commit>=4.3.0",
    "ruff>=0.12.11",
]
